requirements.txt
google-endpoints==3.0.0
google-endpoints-api-management==1.5.1
lxml==3.7.3
